Added somewhat working infrastructure for menus
The Authorization now contains all permissions that a user has. It also
has a check whether correct authorization is available. All of this is
a bit hacky for now and I am looking into a better way to provide an
interface similar to what I have now (without leaking a pure
implementation type class ... and also without building the set of
available permissions that often).

The menu should now have access to all roles; it can link to all API
endpoints. It does not yet automatically determine whether a link shall
be displayed (i.e. the user would have permissions) or not. I think this
should be feasible and also a welcome change in many situations.