Too large of a commit
First, this fixes tests by moving the schemas to the proper location,
where they can be found by the test suite.

Second, this improves test coverage for the Union class and fixes a bug
in the embedding of one union into another that also caused issues with
logging out.

Third, it refactors the backend test suite a bit by moving dummy data
generation into separate modules. It also adds a couple of tests for the
Session effect and improves the Buk.TestUtil.Err module by providing
more details in error message.

Other significant changes are related to the frontend, where we add
another stylesheet and provide a welcome page as well as a basic layout
that is taken from the old implementation.